backend/comments_service/pyproject.toml
[project]
name = "comments-service"
version = "0.1.0"
description = ""
authors = [
    {name = "IvanZaycev0717",email = "ivzaycev0717@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "grpcio (>=1.75.1,<2.0.0)",
    "grpcio-tools (>=1.75.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic[email] (>=2.12.0,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "grpcio-reflection (>=1.75.1,<2.0.0)",
    "grpcio-health-checking (>=1.75.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
